[perf] bound N-scaling memory: chunk generation + stream hidden_states
Full run (N=2500) OOM'd: two N-coupled blowups the N=30 smoke (83x smaller)
could not surface.

- generate: hf_local.generate batched ALL N prompts x n_samples in one
  model.generate => 25000 sequences at N=2500 -> SiLU OOM (~20GB) + decode KV.
  Now chunked by GenerateStage.batch_size (default 0 = old single-batch,
  byte-identical). Full configs use 64 (sampled GPU batch 640 seqs, KV ~11GB;
  256 would OOM on the decode KV cache ~44GB).
- hidden_states: accumulated all N residual tensors in fp32 on CPU (~120-160GB
  host RAM + 81GB fp32 disk). Now a generator streaming one sample at a time in
  a configurable dtype; per-layer slices are .clone()d (a contiguous view would
  keep the whole stacked tensor alive) so del frees it each iteration. Full
  configs: storage=question_only (only last_token_q is pooled) + dtype=bfloat16
  -> ~20GB RAM / ~12GB disk.
- vuf upcasts to fp32 before the diff-in-means so bf16 storage keeps precision.

Defaults (batch_size=0, storage=full, dtype=float32) keep existing/smoke runs
byte-identical. Also records the share-method-independent-upstream TODO (HIGH).

+++ src/sae_muc/pipeline/vuf.py
@@ -215,8 +215,11 @@ def run(ctx: PipelineContext) -> list[str]:
         tensors = ctx.store.load_safetensors(_layer_in_path(layer))
 
         def pooled(sid: str, _tensors=tensors) -> "torch.Tensor":
+            # Upcast to float32 before the diff-in-means so a bf16 hidden-state
+            # artefact (large-N runs) doesn't lose precision when averaging
+            # hundreds of vectors. No-op when the artefact is already float32.
             return _pool(
-                _tensors[sid],
+                _tensors[sid].float(),
                 stage_cfg.pooling,
                 int(meta.loc[sid, "question_len"]),
                 int(meta.loc[sid, "seq_len"]),
